requirements.txt
setuptools~=60.5.0
PyYAML~=6.0